Added the logout functionality and the AuthPage button. Also edited GoogleOAuthLoginBtn to add debugging functionality.

+++ backend/users/views.py
@@ -20,7 +20,6 @@ from users.serializers import UserSerializer
 logger = logging.getLogger(__name__)
 
 
-
 # NOTE: Helper function for setting cookies (consider moving into separate file/Class)
 def set_authentication_cookies(response, access_token, refresh_token, request):
     response.set_cookie(
@@ -44,8 +43,8 @@ def set_authentication_cookies(response, access_token, refresh_token, request):
     return response
 
 
-
 def remove_authenticated_cookies(response):
+    logger.debug('Removing authenticated cookies')
     response.delete_cookie('access_token')
     response.delete_cookie('refresh_token')
     response.delete_cookie('csrftoken')
@@ -58,26 +57,31 @@ def logout_view(request):
     logger.debug("Logout view called")
     try:
         logger.debug("User: %s", request.user)
-        Token.objects.filter(user=request.user).delete()
+        logger.debug("Request headers: %s", request.headers)
+        logger.debug("User is authenticated: %s", request.user.is_authenticated)
+        
+        if isinstance(request.user, AnonymousUser):
+            logger.debug("User is AnonymousUser")
+            return Response({'error': 'User not authenticated'}, status=status.HTTP_401_UNAUTHORIZED)
 
-        #! important logout as django_logout to avoid conflicts
+        logger.debug("User is authenticated and not AnonymousUser")
+        Token.objects.filter(user=request.user).delete()
         django_logout(request)
         response = Response({'message': 'User logged out successfully'}, status=status.HTTP_200_OK)
         response = remove_authenticated_cookies(response)
-        logger.info("%s logged out", request.user.email)
+        logger.debug('Cookies removed')
+
+        if hasattr(request.user, 'email'):
+            logger.info("%s logged out", request.user.email)
+        else:
+            logger.info("User without email attribute logged out")
+
         return response
     except Exception as e:
         logger.error("Logout error: %s", str(e))
         return Response({'error': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
 
-
-
-
-# NOTE: Currently somewhat of a hack using python's native requests lib,
-# there is probably a better way using the social_auth.core
-# TODO: This route's logic is getting too long,
-# refactor into smaller helper functions
 @api_view(['POST'])
 @permission_classes([AllowAny])
 @psa()
@@ -88,7 +92,6 @@ def register_by_access_token(request, backend):
         return Response({'error': 'No OAuth code provided.'},
                         status=status.HTTP_400_BAD_REQUEST)
     try:
-        # NOTE: Consider moving these settings into separate file/Class
         token_url = 'https://www.googleapis.com/oauth2/v4/token'
         data = {
             'code': code,
@@ -106,30 +109,21 @@ def register_by_access_token(request, backend):
 
         tokens = response.json()
         access_token = tokens.get('access_token')
-        # TODO: refresh_token should be set in key/value store (i.e. redis cache)
-        # and used to grab new access_token when access_token is expired
-        # TODO: refresh_token could not exist, handle exception, but don't return
         refresh_token = tokens.get('refresh_token')
         user = request.backend.do_auth(access_token)
 
         if user:
-            # NOTE: Temporary demonstration of "sign up" logic,
-            # Uses Django REST framework serializers to save user to DB
-            # TODO: We Need an Onboarding Process that establishes
-            # the user_name (user chooses username)
             user_data = {
                 'user_name': user.email,
                 'user_email': user.email,
             }
 
             serializer = UserSerializer(data=user_data)
-            existing_user = serializer.get_user_by_email(  # type:ignore
-                data=user_data)
+            existing_user = serializer.get_user_by_email(data=user_data)  # type:ignore
             if existing_user is not None:
                 res = Response(
                     {
-                        'message':
-                        'Sign Up Failed. User Already Exists. Please Login.'
+                        'message': 'Sign Up Failed. User Already Exists. Please Login.'
                     },
                     status=status.HTTP_409_CONFLICT)
                 logger.warning(
@@ -140,15 +134,16 @@ def register_by_access_token(request, backend):
                 serializer.save()
 
             token, _ = Token.objects.get_or_create(user=user)  # type:ignore
-            # NOTE: Sets sessionid cookie, is necessary??
-            # TODO: Investigate Django login() method more
             login(request, user)
             res = Response(
-                {'message': 'User Authenticated, setting credentials'},
+                {
+                    'message': 'User Authenticated, setting credentials',
+                    'access_token': access_token,
+                    'refresh_token': refresh_token,
+                },
                 status=status.HTTP_200_OK,
             )
-            res = set_authentication_cookies(res, token.key, refresh_token,
-                                             request)
+            res = set_authentication_cookies(res, token.key, refresh_token, request)
 
             logger.info("%s successful signed up", user.email)
             return res
@@ -164,8 +159,6 @@ def register_by_access_token(request, backend):
                         status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
 
-# TODO: This route's logic is getting too long,
-# refactor into smaller helper functions
 @api_view(['POST'])
 @permission_classes([AllowAny])
 @psa()
@@ -176,7 +169,6 @@ def login_by_access_token(request, backend):
         return Response({'error': 'No OAuth code provided.'},
                         status=status.HTTP_400_BAD_REQUEST)
     try:
-        # NOTE: Consider moving these settings into separate file/Class
         token_url = 'https://www.googleapis.com/oauth2/v4/token'
         data = {
             'code': code,
@@ -204,13 +196,11 @@ def login_by_access_token(request, backend):
             }
 
             serializer = UserSerializer(data=user_data)
-            existing_user = serializer.get_user_by_email(  # type:ignore
-                data=user_data)
+            existing_user = serializer.get_user_by_email(data=user_data)  # type:ignore
             if existing_user is None:
                 res = Response(
                     {
-                        'message':
-                        'Login Failed. User Does Not Exist. Please Sign Up.'
+                        'message': 'Login Failed. User Does Not Exist. Please Sign Up.'
                     },
                     status=status.HTTP_401_UNAUTHORIZED)
                 logger.warning(
@@ -222,13 +212,16 @@ def login_by_access_token(request, backend):
             login(request, user)
 
             res = Response(
-                {'message': 'User Is Authenticated, Logging In...'},
+                {
+                    'message': 'User Is Authenticated, Logging In...',
+                    'access_token': access_token,
+                    'refresh_token': refresh_token,
+                },
                 status=status.HTTP_200_OK,
             )
-            res = set_authentication_cookies(res, token.key, refresh_token,
-                                             request)
+            res = set_authentication_cookies(res, token.key, refresh_token, request)
 
-            logger.info("%s is successfully authenticated, loggging in...",
+            logger.info("%s is successfully authenticated, logging in...",
                         user.email)
             return res
         else:
@@ -243,8 +236,6 @@ def login_by_access_token(request, backend):
                         status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
 
-# TODO: redirect to get new access_token if access_token is expired via refresh_token
-# TODO: Have this check the Users DB to see if they exist in the DB every time...
 @api_view(['POST'])
 @csrf_protect
 def authentication_test(request):
@@ -254,7 +245,7 @@ def authentication_test(request):
             {"message": "Access forbidden. You are not authenticated."},
             status=status.HTTP_401_UNAUTHORIZED)
 
-    logger.info("%s successful authenticated", request.user)
+    logger.info("%s successfully authenticated", request.user)
     return Response(
         {'message': "User successfully authenticated"},
         status=status.HTTP_200_OK,
